JS errors thrown in private scripts should be reported to stderr in Debug builds
In order to debug private scripts, JS errors thrown in the private scripts
(e.g., ReferenceError thrown when we make a typo in private scripts)
need to be reported to stderr. This CL makes the change.

The fprintf(stderr, ...) is enabled only in Debug builds.

This CL also renames throwDOMExceptionInPrivateScriptIfNeeded to rethrowExceptionInPrivateScript.

BUG=341031

Review URL: https://codereview.chromium.org/460573002

git-svn-id: svn://svn.chromium.org/blink/trunk@180033 bbb929c8-8fbe-4397-9dbb-9b2b20218538