Add lextab.pyc as an output of the cached_lex_yacc_tables action
The primary output is lextab.py, but since it is imported, Python also
writes lextab.pyc. Listing it too as an output means it won't be left
around by "ninja -t clean" (and not much else, I think.)

Prior to the patch
  https://codereview.chromium.org/425953002/
the result of not having lextab.pyc listed as an output and running
"ninja -t clean" was a tree that would build correctly, but that would
always rerun all the IDL code generation scripts, and thus never become
"clean".

BUG=397909

Review URL: https://codereview.chromium.org/463063003

git-svn-id: svn://svn.chromium.org/blink/trunk@180142 bbb929c8-8fbe-4397-9dbb-9b2b20218538