Extract "fullscreen element ready check" from requestFullscreen()
http://fullscreen.spec.whatwg.org/#fullscreen-element-ready-check

This is a change in behavior for what was previously "a descendant
browsing context's document has a non-empty fullscreen element stack,"
now "element has no ancestor element whose local name is iframe and
namespace is the HTML namespace."

It's now never possible to go fullscreen with a descendant of an iframe,
whereas previously it would be possible as long as that iframe's
document had an empty fullscreen element stack.

Since the difference is very unlikely to affect any real content, the
LegacyFullScreenErrorExemption was dropped for this case. There was
apparently no test coverage for that exemption.

BUG=402376

Review URL: https://codereview.chromium.org/462103002

git-svn-id: svn://svn.chromium.org/blink/trunk@180162 bbb929c8-8fbe-4397-9dbb-9b2b20218538