Oilpan: fix tracing of un-initialized part objects during conservative GCs.
If a part object has a virtual trace method we need to check the vtable
before calling the trace method. This is necessary because conservative
GCs can see the part object before it has been constructed (if there
is an allocation during the construction of the containing object.)

Added simple regression test that illustrates the issue.

R=erik.corry@gmail.com, zerny@chromium.org

Review URL: https://codereview.chromium.org/455363002

git-svn-id: svn://svn.chromium.org/blink/trunk@180335 bbb929c8-8fbe-4397-9dbb-9b2b20218538