Simplify watched selector handling and remove remnants of user stylesheets
This patch simplifies our handling of watched selectors, allowing us to
remove the remaining pieces of user stylesheets. Currently a special
-internal-callback property is added via a user stylesheet, which depends
on the StyleBuilder applying all values and not just the cascaded values
to function correctly. The approach in this patch is to have this logic
separate to applyMatchedProperties, which means we won't need to have a
separate property and we can still use the matched properties cache in
these cases. This also allows us to remove the Vector of StyleRule in
MatchResult.

BUG=316960

Review URL: https://codereview.chromium.org/418163003

git-svn-id: svn://svn.chromium.org/blink/trunk@180379 bbb929c8-8fbe-4397-9dbb-9b2b20218538