Fix crash on GestureTap in Node::commonAncestor
When doing a tap, if the synthetic mousedown lands on some node, but
changes the DOM such that there's no node at that point afterwards
(eg. a frame scrollbar), then we'll pass a null node pointer to
commonAncestor and crash.

In addition to adding a test for this case (the mousedown case in
gesture-tap-frame-scrollbar) I added tests for a number of similar
cases (null hit test and removed node) that are already working
correctly but could be subject to similar bugs.

Also corrected a minor typo in a related test
(gesture-tap-click-common-ancestor) - doesn't change behavior, just
better reflects the intention of the test (the ancestor node itself
should not be hittable).

BUG=402338

Review URL: https://codereview.chromium.org/466143005

git-svn-id: svn://svn.chromium.org/blink/trunk@180389 bbb929c8-8fbe-4397-9dbb-9b2b20218538