Don't use ServiceWorkerContainer.unregister in service_worker_unregister_and_register
Another step toward removing the deprecated ServiceWorkerContainer.unregister.
service_worker_unregister_and_register is the biggest consumer of it in our
tests.

Now to unregister, the function first registers at that scope to get a
registration, then calls unregister on it. This only works with registrations
that cleanly install, which is the most common use case.

BUG=404064

Review URL: https://codereview.chromium.org/472293002

git-svn-id: svn://svn.chromium.org/blink/trunk@180426 bbb929c8-8fbe-4397-9dbb-9b2b20218538