All selectors in :-webkit-any must have an invalidation set feature.
In order to avoid subtree style recalc for a :-webkit-any selector in the
rightmost compound selector, all of its selectors must contain an
invalidation set feature. In terms of invalidation sets, these selectors
are all universal in the rightmost compound:

.x :-webkit-any(*)
.x :-webkit-any(.y, *)
.x :-webkit-any(.y, :hover)

The fix is to set foundIdent to true when all selectors in the selector
list have found one.

BUG=392852

Review URL: https://codereview.chromium.org/383833002

git-svn-id: svn://svn.chromium.org/blink/trunk@180566 bbb929c8-8fbe-4397-9dbb-9b2b20218538