Don't allow setting properties on scriptable v8 plugin objects
Plugin elements can be backed by v8 objects which are used for
scripting from JS into the plugin. There are no longer any use cases
where we need to allow setting properties on the backing object as a
result of setting the property on the plugin element. As a result we
can remove the call to the backing v8 object.

Note that this should only affect Pepper plugins that support
scripting (Flash and PDF being the ones of importance) but I have
verified that they do not need this. When NPAPI is removed, we should
be able to remove the corresponding NPObject codepath.

BUG=351636

Review URL: https://codereview.chromium.org/472243002

git-svn-id: svn://svn.chromium.org/blink/trunk@180581 bbb929c8-8fbe-4397-9dbb-9b2b20218538