Credential Manager: Migrate to WebView client-based approach.
In [1], jam@ suggested that we change the current approach of delivering
the core credential manager implementation via a Platform API piped
through //content. Instead, he proposed creating a client interface, and
setting that directly on the WebView from ChromeContentRendererClient.
This bypasses //content entirely, and makes the Chromium-side
implementation cleaner.

The Blink-side implementation gets a bit messier. In order to ensure
that credential manager doesn't introduce changes to core, this patch
introduces a new client interface in the module, and supplements Page
in order to make this interface available to the module's bindings code.
This is similar to the approach PrerendererClient has taken.

[1]: https://codereview.chromium.org/464883002/#msg9

BUG=400674

Review URL: https://codereview.chromium.org/488993003

git-svn-id: svn://svn.chromium.org/blink/trunk@180648 bbb929c8-8fbe-4397-9dbb-9b2b20218538