Add checking queue size when calling handleSpeakingCompleted().
Some speech contents are not removed when is called by async.
In this case, when call to speak() function twice, second data is not removed in queue.

So I add routine in handleSpeakingCompleted() that is start to speak() function immediately when latest queue size is not 0.
It also works fine if called by asynchronous repeatedly.
BUG=405585

Review URL: https://codereview.chromium.org/495603002

git-svn-id: svn://svn.chromium.org/blink/trunk@180813 bbb929c8-8fbe-4397-9dbb-9b2b20218538