Oilpan: Promptly free collection backings when possible.
Because collection backings are owned by their front object
they often have a known destruction time (for example when
growing the backing size). A collection backing can be freed
immediately so long as a GC is not currently in progress.
Freeing the backing allows reuse of the backing memory and
may postpone the need to do a full GC.

R=ager@chromium.org
BUG=

Review URL: https://codereview.chromium.org/493423002

git-svn-id: svn://svn.chromium.org/blink/trunk@180957 bbb929c8-8fbe-4397-9dbb-9b2b20218538