Add [NoImplHeader] extended attribute
If all of attributes and methods of an interface are implemented in private scripts,
there is no header file in the C++ side. To prevent the IDL compiler from including
non-existing header files, this CL adds an IDL extended attribute to specify the fact.

BUG=341031

Review URL: https://codereview.chromium.org/519703002

git-svn-id: svn://svn.chromium.org/blink/trunk@181093 bbb929c8-8fbe-4397-9dbb-9b2b20218538