Oilpan: Move Panner and HRTFDatabaseLoader to oilpan's heap
This is one of the leftovers when we shipped Oilpan for webaudio/.

The only substantial change of this CL is to change HRTFDatabaseLoader::LoaderMap to a static persistent heap hash map of weak members. By doing this, we can remove map.remove() from HRTFDatabaseLoader's destructor.

BUG=340522

Review URL: https://codereview.chromium.org/530723004

git-svn-id: svn://svn.chromium.org/blink/trunk@181211 bbb929c8-8fbe-4397-9dbb-9b2b20218538