Modify PageWidgetDelegate methods to work with non-main frames.
WebWidget will soon be able to be associated with multiple frames on
a single page, and accordingly needs the ability to layout and paint
on subtrees of the frame tree. This change had already been made for
layout but the other PageWidgetDelegate methods need it as well.

As with layout, the methods are receiving a default LocalFrame argument
which, when NULL, causes PageWidgetDelegate to attempt to use the
Page mainFrame(), matching existing behavior.

BUG=346764

Review URL: https://codereview.chromium.org/514253003

git-svn-id: svn://svn.chromium.org/blink/trunk@181245 bbb929c8-8fbe-4397-9dbb-9b2b20218538